+++ packages/core/src/compile.test.ts
@@ -1,6 +1,8 @@
 import { describe, it } from 'vitest';
 import { createFixture } from 'fs-fixture';
 import { compile } from './compile';
+import { mkdir, writeFile } from 'node:fs/promises';
+import { join } from 'node:path';
 import type { UserConfig } from './types';
 
 const wait = (ms: number) => new Promise((resolve) => setTimeout(resolve, ms));
@@ -22,7 +24,6 @@ const DEFAULT_FIXTURE = {
 const baseConfig: UserConfig = {
   devLanguage: 'en',
   languages: [{ name: 'en' }, { name: 'fr' }],
-  projectRoot: '',
 };
 
 // Wrapper to make fixture compatible with `await using`
@@ -181,6 +182,80 @@ describe.concurrent('compile', () => {
       await expect(
         fixture.readFile('src/.ignored/.vocab/index.ts'),
       ).rejects.toThrow();
+    });
+
+    it('should discover new translation directories', async ({ expect }) => {
+      await using fixture = await createVocabFixture(DEFAULT_FIXTURE);
+      const stopWatching = await compile(
+        { watch: true },
+        { ...baseConfig, projectRoot: fixture.path },
+      );
+      await wait(500);
+
+      // Manually create new translation directory and file because fs-fixture doesn't give us
+      // methods to do that
+      await mkdir(join(fixture.path, 'src/new.vocab/'), {
+        recursive: true,
+      });
+      await writeFile(
+        join(fixture.path, 'src/new.vocab/translations.json'),
+        JSON.stringify({
+          test: { message: 'New file' },
+        }),
+      );
+      await wait(500);
+
+      const newCompiledFile = await fixture.readFile(
+        'src/new.vocab/index.ts',
+        'utf-8',
+      );
+      expect(newCompiledFile).toMatchSnapshot();
+
+      await fixture.writeFile(
+        'src/new.vocab/translations.json',
+        JSON.stringify({
+          test: { message: 'Changed new file' },
+        }),
+      );
+      await wait(500);
+      const updatedNewCompiledFile = await fixture.readFile(
+        'src/new.vocab/index.ts',
+        'utf-8',
+      );
+      expect(updatedNewCompiledFile).toMatchSnapshot();
+
+      await stopWatching?.();
+    });
+
+    it('should create and update index.ts when translation file path starts with "."', async ({
+      expect,
+    }) => {
+      await using fixture = await createVocabFixture(DEFAULT_FIXTURE);
+      const stopWatching = await compile(
+        { watch: true },
+        { ...baseConfig, projectRoot: join(fixture.path, 'src') },
+      );
+      await wait(500);
+
+      await expect(
+        fixture.readFile('src/.vocab/index.ts', 'utf-8'),
+      ).resolves.toBeTruthy();
+
+      await fixture.writeFile(
+        'src/.vocab/translations.json',
+        JSON.stringify({
+          hello: { message: 'Howdy', description: 'A greeting' },
+          world: { message: 'World' },
+        }),
+      );
+      await wait(500);
+      const indexContent = await fixture.readFile(
+        'src/.vocab/index.ts',
+        'utf-8',
+      );
+
+      expect(indexContent).toMatchSnapshot();
+      await stopWatching?.();
     });
   });