+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @vocab/cli
 
+## 2.1.9
+
+### Patch Changes
+
+- Updated dependencies [[`5f5bdb5`](https://github.com/seek-oss/vocab/commit/5f5bdb5e323c0121491d923dcaae961c5ac3fe81)]:
+  - @vocab/core@1.7.5
+  - @vocab/phrase@2.1.9
+
+## 2.1.8
+
+### Patch Changes
+
+- Updated dependencies [[`710649f`](https://github.com/seek-oss/vocab/commit/710649fae726227701f51410d6050fd0c17d6e3f)]:
+  - @vocab/core@1.7.4
+  - @vocab/phrase@2.1.8
+
 ## 2.1.7
 
 ### Patch Changes