+++ packages/react/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @vocab/react
 
+## 1.1.21
+
+### Patch Changes
+
+- Updated dependencies [[`5f5bdb5`](https://github.com/seek-oss/vocab/commit/5f5bdb5e323c0121491d923dcaae961c5ac3fe81)]:
+  - @vocab/core@1.7.5
+
+## 1.1.20
+
+### Patch Changes
+
+- Updated dependencies [[`710649f`](https://github.com/seek-oss/vocab/commit/710649fae726227701f51410d6050fd0c17d6e3f)]:
+  - @vocab/core@1.7.4
+
 ## 1.1.19
 
 ### Patch Changes